+++ packages/capture-main/src/core/getCrossOriginWorkerURL.ts
@@ -40,11 +40,32 @@ export const getCrossOriginWorkerURL = (
     return Promise.resolve(originalWorkerUrl);
   }
 
+  let signal: AbortSignal;
+
+  try {
+    // Setup the abort controller
+    const controller = new AbortController();
+    signal = controller.signal;
+
+    const timeout = setTimeout(() => {
+      controller.abort();
+    }, 3000);
+
+    const cleanup = () => {
+      clearTimeout(timeout);
+      controller.abort();
+    };
+
+    signal.addEventListener("abort", cleanup);
+  } catch (error) {
+    // just swallow the error
+  }
+
   return new Promise<string>(
     (resolve, reject) =>
       void fetch(originalWorkerUrl, {
         // abort if the worker is not fetched in a reasonable time
-        signal: AbortSignal.timeout(3000),
+        signal,
       })
         .then((res) => res.text())
         .then((codeString) => {