+++ packages/capture-main/CHANGELOG.md
@@ -1,5 +1,12 @@
 # capture-main
 
+## 1.2.2
+
+### Patch Changes
+
+- capture-wasm@1.2.2
+- capture-worker@1.2.2
+
 ## 1.2.1
 
 ### Patch Changes